Move classes to internal packages
