+++ bundles/org.corpus_tools.hexatomic.grid/src/main/java/org/corpus_tools/hexatomic/grid/GridEditor.java
@@ -30,20 +30,20 @@ import org.corpus_tools.hexatomic.core.SaltHelper;
 import org.corpus_tools.hexatomic.core.Topics;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
 import org.corpus_tools.hexatomic.core.handlers.OpenSaltDocumentHandler;
-import org.corpus_tools.hexatomic.grid.bindings.FreezeGridBindings;
-import org.corpus_tools.hexatomic.grid.configuration.BodyMenuConfiguration;
-import org.corpus_tools.hexatomic.grid.configuration.ColumnHeaderMenuConfiguration;
-import org.corpus_tools.hexatomic.grid.configuration.ColumnHeaderUiBindingConfiguration;
-import org.corpus_tools.hexatomic.grid.configuration.EditConfiguration;
-import org.corpus_tools.hexatomic.grid.configuration.GridLayerConfiguration;
-import org.corpus_tools.hexatomic.grid.data.ColumnHeaderDataProvider;
-import org.corpus_tools.hexatomic.grid.data.GraphDataProvider;
-import org.corpus_tools.hexatomic.grid.data.NodeSpanningDataProvider;
-import org.corpus_tools.hexatomic.grid.data.RowHeaderDataProvider;
-import org.corpus_tools.hexatomic.grid.layers.GridColumnHeaderLayer;
-import org.corpus_tools.hexatomic.grid.layers.GridSpanningDataLayer;
-import org.corpus_tools.hexatomic.grid.style.LabelAccumulator;
-import org.corpus_tools.hexatomic.grid.style.SelectionStyleConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.bindings.FreezeGridBindings;
+import org.corpus_tools.hexatomic.grid.internal.configuration.BodyMenuConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.configuration.ColumnHeaderMenuConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.configuration.ColumnHeaderUiBindingConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.configuration.EditConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.configuration.GridLayerConfiguration;
+import org.corpus_tools.hexatomic.grid.internal.data.ColumnHeaderDataProvider;
+import org.corpus_tools.hexatomic.grid.internal.data.GraphDataProvider;
+import org.corpus_tools.hexatomic.grid.internal.data.NodeSpanningDataProvider;
+import org.corpus_tools.hexatomic.grid.internal.data.RowHeaderDataProvider;
+import org.corpus_tools.hexatomic.grid.internal.layers.GridColumnHeaderLayer;
+import org.corpus_tools.hexatomic.grid.internal.layers.GridSpanningDataLayer;
+import org.corpus_tools.hexatomic.grid.internal.style.LabelAccumulator;
+import org.corpus_tools.hexatomic.grid.internal.style.SelectionStyleConfiguration;
 import org.corpus_tools.hexatomic.grid.style.StyleConfiguration;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
