+++ bundles/org.corpus_tools.hexatomic.grid/src/main/java/org/corpus_tools/hexatomic/grid/internal/handlers/DisplayAnnotationRenameDialogCommandHandler.java
@@ -1,0 +1,64 @@
+/*-
+ * #%L
+ * org.corpus_tools.hexatomic.grid
+ * %%
+ * Copyright (C) 2018 - 2020 Stephan Druskat, Thomas Krause
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+package org.corpus_tools.hexatomic.grid.internal.handlers;
+
+import org.corpus_tools.hexatomic.grid.internal.commands.DisplayAnnotationRenameDialogCommand;
+import org.corpus_tools.hexatomic.grid.internal.commands.RenameCellAnnotationCommand;
+import org.corpus_tools.hexatomic.grid.internal.ui.AnnotationRenameDialog;
+import org.eclipse.nebula.widgets.nattable.command.AbstractLayerCommandHandler;
+import org.eclipse.swt.widgets.Display;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Displays an {@link AnnotationRenameDialog} on selected cells.
+ * 
+ * @author Stephan Druskat {@literal <mail@sdruskat.net>}
+ */
+public class DisplayAnnotationRenameDialogCommandHandler
+    extends AbstractLayerCommandHandler<DisplayAnnotationRenameDialogCommand> {
+
+  private static final Logger log =
+      LoggerFactory.getLogger(DisplayAnnotationRenameDialogCommandHandler.class);
+
+  @Override
+  public Class<DisplayAnnotationRenameDialogCommand> getCommandClass() {
+    return DisplayAnnotationRenameDialogCommand.class;
+  }
+
+  @Override
+  protected boolean doCommand(DisplayAnnotationRenameDialogCommand command) {
+    log.debug("Executing command {}.", getCommandClass().getSimpleName());
+    AnnotationRenameDialog dialog =
+        new AnnotationRenameDialog(Display.getDefault().getActiveShell(), null, null);
+    dialog.open();
+
+    if (dialog.isCancelPressed()) {
+      log.debug("Execution of command {} cancelled.", getCommandClass().getSimpleName());
+      return true;
+    }
+
+    log.debug("Returning delegate command {}.", RenameCellAnnotationCommand.class.getSimpleName());
+    return command.getNatTable().doCommand(
+        new RenameCellAnnotationCommand(command.getSelectedCellPositions(), command.getNatTable()));
+  }
+
+}
