+++ bundles/org.corpus_tools.hexatomic.grid/src/main/java/org/corpus_tools/hexatomic/grid/internal/handlers/RenameCellAnnotationCommandHandler.java
@@ -1,0 +1,50 @@
+/*-
+ * #%L
+ * org.corpus_tools.hexatomic.grid
+ * %%
+ * Copyright (C) 2018 - 2020 Stephan Druskat, Thomas Krause
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+package org.corpus_tools.hexatomic.grid.internal.handlers;
+
+import org.corpus_tools.hexatomic.grid.internal.commands.RenameCellAnnotationCommand;
+import org.eclipse.nebula.widgets.nattable.command.AbstractLayerCommandHandler;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * A handler for the {@link RenameCellAnnotationCommand}.
+ * 
+ * @author Stephan Druskat {@literal <mail@sdruskat.net>}
+ */
+public class RenameCellAnnotationCommandHandler
+    extends AbstractLayerCommandHandler<RenameCellAnnotationCommand> {
+
+  private static final Logger log =
+      LoggerFactory.getLogger(RenameCellAnnotationCommandHandler.class);
+
+  @Override
+  public Class<RenameCellAnnotationCommand> getCommandClass() {
+    return RenameCellAnnotationCommand.class;
+  }
+
+  @Override
+  protected boolean doCommand(RenameCellAnnotationCommand command) {
+    log.debug("RUNNING THE COMMAND!");
+    return true;
+  }
+
+}
